working with refactor
